test: refactor plugin utility position
