+++ packages/acceptance-tests/pkg-tests-specs/sources/features/plugins.test.ts
@@ -1,36 +1,9 @@
 import {hashUtils}         from '@yarnpkg/core';
 import {PortablePath, xfs} from '@yarnpkg/fslib';
 import {stringifySyml}     from '@yarnpkg/parsers';
-import https               from 'https';
-import {AddressInfo}       from 'net';
-import {tests, fs}         from 'pkg-tests-core';
+import {fs}                from 'pkg-tests-core';
 
-const mockPluginServer: (path: PortablePath) => Promise<{pluginUrl: string, httpsCaFilePath: PortablePath}> = async path => {
-  return new Promise((resolve, reject) => {
-    (async () => {
-      const certs = await tests.getHttpsCertificates();
-      const httpsCaFilePath = `${path}/rootCA.crt` as PortablePath;
-      await xfs.writeFilePromise(httpsCaFilePath, certs.ca.certificate);
-
-      const helloWorldPluginPath = require.resolve(`@yarnpkg/monorepo/scripts/plugin-hello-world.js`) as PortablePath;
-      const helloWorldPlugin = await xfs.readFilePromise(helloWorldPluginPath);
-      const server = https.createServer({
-        cert: certs.server.certificate,
-        key: certs.server.clientKey,
-        ca: certs.ca.certificate,
-      }, (req, res) => {
-        res.writeHead(200);
-        res.end(helloWorldPlugin);
-      });
-
-      server.unref();
-      server.listen(() => {
-        const {port} = server.address() as AddressInfo;
-        resolve({pluginUrl: `https://localhost:${port}`, httpsCaFilePath});
-      });
-    })();
-  });
-};
+import {mockPluginServer}  from './plugins.utility';
 
 const PLUGIN = (name: string, {async = false, printOnBoot = false} = {}) => `
 const factory = ${async ? `async` : ``} r => {
